Handle the case when month has taken 42 cells to render
If month has around 31 days and it started at the end of the week
our default 35 cells are not enough and we should render one more
row to show last days of the month.

Replace render logic for cells.
Now we are creating cells in memory inside Fragment and then
replace the whole body of month at a time.